Move plugins to own directory
One should not put plugins into libraries. Instead, plugins should
be built directly into shared plugin objects.
This fixes the problem in the IB where L1GlobalTrigger module could
not be found. It could not be found since the library it was contained
within was no longer being indirectly linked to some othe plugin.
