+++ L1Trigger/GlobalTrigger/plugins/BuildFile.xml
@@ -1,0 +1,5 @@
+<export>
+</export>
+<library name="L1TriggerGlobalTriggerPlugins" file="*.cc">
+  <use name="L1Trigger/GlobalTrigger"/>
+</library>
