+++ L1Trigger/GlobalTrigger/plugins/CompareToObjectMapRecord.h
