+++ L1Trigger/GlobalTrigger/plugins/ConvertObjectMapRecord.h
