+++ L1Trigger/GlobalTrigger/plugins/L1GlobalTrigger.h
