--- L1Trigger/GlobalTrigger/src/SealModule.cc
@@ -1,8 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "L1Trigger/GlobalTrigger/interface/L1GlobalTrigger.h"
-#include "L1Trigger/GlobalTrigger/interface/ConvertObjectMapRecord.h"
-#include "L1Trigger/GlobalTrigger/interface/CompareToObjectMapRecord.h"
-  
-DEFINE_FWK_MODULE(L1GlobalTrigger);
-DEFINE_FWK_MODULE(ConvertObjectMapRecord);
-DEFINE_FWK_MODULE(CompareToObjectMapRecord);
